feat: create aave-like open multiply strategy
