+++ packages/dma-library/src/protocols/aave-like/index.ts
@@ -1,0 +1,1 @@
+export { isAaveProtocol, resolveAavelikeProtocol } from './resolve-aavelike-protocols'
