+++ packages/dma-library/src/protocols/aave/get-aave-protocol-data.ts
@@ -1,5 +1,8 @@
-import { AaveLikeStrategyAddresses } from '@dma-library/operations/aave-like'
-import { PriceResult, ReserveDataResult } from '@dma-library/protocols/aave-like/types'
+import {
+  PriceResult,
+  ReserveDataResult,
+  SharedAaveLikeProtocolDataArgs,
+} from '@dma-library/protocols/aave-like/types'
 import {
   determineReserveEModeCategory,
   fetchAssetPrice,
@@ -9,21 +12,11 @@ import {
 } from '@dma-library/protocols/aave-like/utils'
 import * as AaveCommon from '@dma-library/strategies/aave/common'
 import { AaveVersion } from '@dma-library/types/aave'
-import { providers } from 'ethers'
 
-type SharedAaveProtocolDataArgs = {
-  collateralTokenAddress: string
-  debtTokenAddress: string
-  addresses: AaveLikeStrategyAddresses
-  provider: providers.Provider
-  flashloanTokenAddress?: string
-  proxy?: string
-}
-
-export type AaveV2ProtocolDataArgs = SharedAaveProtocolDataArgs & {
+export type AaveV2ProtocolDataArgs = SharedAaveLikeProtocolDataArgs & {
   protocolVersion: AaveVersion.v2
 }
-export type AaveV3ProtocolDataArgs = SharedAaveProtocolDataArgs & {
+export type AaveV3ProtocolDataArgs = SharedAaveLikeProtocolDataArgs & {
   protocolVersion: AaveVersion.v3
 }
 export type AaveProtocolDataArgs = AaveV2ProtocolDataArgs | AaveV3ProtocolDataArgs
@@ -46,7 +39,7 @@ export const getAaveProtocolData: GetAaveProtocolData = async args => {
   if (
     AaveCommon.isV2<
       AaveProtocolDataArgs,
-      SharedAaveProtocolDataArgs & { protocolVersion: AaveVersion.v2 }
+      SharedAaveLikeProtocolDataArgs & { protocolVersion: AaveVersion.v2 }
     >(args)
   ) {
     return getAaveV2ProtocolData(args)
@@ -106,7 +99,7 @@ export async function getAaveV3ProtocolData({
   collateralTokenAddress,
   flashloanTokenAddress,
   proxy,
-}: SharedAaveProtocolDataArgs & { protocolVersion: AaveVersion.v3 }) {
+}: SharedAaveLikeProtocolDataArgs & { protocolVersion: AaveVersion.v3 }) {
   const { oracle, poolDataProvider, pool } = await getAaveLikeSystemContracts(
     addresses,
     provider,
