+++ packages/dma-library/src/strategies/aave-like/common/index.ts
@@ -1,1 +1,2 @@
 export { resolveCurrentPositionForProtocol } from './resolve-position'
+export { resolveProtocolData } from './resolve-protocol-data'
