+++ packages/dma-library/src/strategies/aave-like/index.ts
@@ -1,12 +1,16 @@
 import { AaveLikeDepositBorrow, depositBorrow } from './borrow/deposit-borrow'
 import { AaveLikeOpenDepositBorrow, openDepositBorrow } from './borrow/open-deposit-borrow'
 import { AaveLikePaybackWithdraw, paybackWithdraw } from './borrow/payback-withdraw'
+import { AaveLikeOpen, open } from './multiply/open'
 
 type AaveLike = {
   borrow: {
     depositBorrow: AaveLikeDepositBorrow
     openDepositBorrow: AaveLikeOpenDepositBorrow
     paybackWithdraw: AaveLikePaybackWithdraw
+  }
+  multiply: {
+    open: AaveLikeOpen
   }
 }
 
@@ -15,5 +19,8 @@ export const aaveLike: AaveLike = {
     depositBorrow,
     openDepositBorrow,
     paybackWithdraw,
+  },
+  multiply: {
+    open,
   },
 }
