+++ packages/dma-library/src/strategies/aave-like/multiply/open/index.ts
@@ -1,2 +1,2 @@
-export { AaveOpenDependencies, AaveV2OpenDependencies, AaveV3OpenDependencies, open } from './open'
-export { AaveOpenArgs } from './types'
+export { open } from './open'
+export { AaveLikeOpen, AaveLikeOpenArgs, AaveLikeOpenDependencies } from './types'
